@@ -8,7 +8,7 @@ build-backend = "setuptools.build_meta"
 
 [tool.black]
 line-length = 88
-target_version = ['py38', 'py39', 'py310']
+target_version = ['py310', 'py311']
 preview = true
 # Exclude irrelevant directories for formatting
 exclude = '''
