@@ -13,9 +13,8 @@ classifiers =
     Intended Audience :: Science/Research
     License :: OSI Approved :: BSD License
     Operating System :: OS Independent
-    Programming Language :: Python :: 3.8
-    Programming Language :: Python :: 3.9
     Programming Language :: Python :: 3.10
+    Programming Language :: Python :: 3.11
     Topic :: Scientific/Engineering
     Topic :: Software Development :: Libraries
 project_urls =
@@ -26,14 +25,12 @@ project_urls =
 include_package_data = True
 packages = find:
 install_requires =
-    scikit-learn>=0.23.0
-    numpy>=1.17.3
-    scipy>=1.4.0
-    pandas>=1.2.0
-    requests
-    joblib
+    scikit-learn>=1.2.1
+    numpy>=1.23.5
+    scipy>=1.9.3
+    pandas>=1.5.3
     pyarrow
-python_requires = >=3.8
+python_requires = >=3.10
 
 [options.extras_require]
 dev =
@@ -46,10 +43,12 @@ dev =
     numpydoc
     flake8
     openml
+    pre-commit
 doc =
     sphinx-gallery
     pydata-sphinx-theme
     sphinxext-opengraph
+    sphinx-copybutton
     matplotlib
     seaborn
     statsmodels
@@ -60,23 +59,15 @@ benchmarks =
     matplotlib
     seaborn
     tqdm
+    thefuzz
+    autofj
 # Overwrite the previous install_requires for CI testing purposes
 # as defined in testing.yml.
-min-py38 =
-    scikit-learn==0.23.0
-    numpy==1.17.3
-    scipy==1.4.0
-    pandas==1.2.0
-min-py39 =
-    scikit-learn==0.24.0
-    numpy==1.19.3
-    scipy==1.6.0
-    pandas==1.2.0
 min-py310 =
-    scikit-learn==1.0.2
-    numpy==1.21.3
-    scipy==1.8.0
-    pandas==1.3.5
+    scikit-learn==1.2.1
+    numpy==1.23.5
+    scipy==1.9.3
+    pandas==1.5.3
 
 [flake8]
 # max line length for black
