@@ -1,5 +1,16 @@
 from pathlib import Path
-from typing import List
+
+from skrub.datasets import (
+    fetch_open_payments,
+    fetch_drug_directory,
+    fetch_road_safety,
+    fetch_midwest_survey,
+    fetch_medical_charge,
+    fetch_employee_salaries,
+    fetch_traffic_violations,
+)
+
+from skrub.datasets._fetching import DatasetAll
 
 import pandas as pd
 
@@ -8,7 +19,7 @@ def find_result(bench_name: str) -> Path:
     return choose_file(find_results(bench_name))
 
 
-def find_results(bench_name: str) -> List[Path]:
+def find_results(bench_name: str) -> list[Path]:
     """
     Returns the list of results in the results' directory.
     """
@@ -16,11 +27,11 @@ def find_results(bench_name: str) -> List[Path]:
     return [
         file
         for file in results_dir.iterdir()
-        if file.stem.startswith(bench_name) and file.suffix == ".csv"
+        if file.stem.startswith(bench_name) and file.suffix == ".parquet"
     ]
 
 
-def choose_file(results: List[Path]) -> Path:
+def choose_file(results: list[Path]) -> Path:
     """
     Given a list of files, chooses one based on these rules:
     - If there are no files to choose from, exit the program
@@ -35,7 +46,7 @@ def choose_file(results: List[Path]) -> Path:
     else:
         for i, file in enumerate(results):
             # Read the result file to get its dimensions
-            df = pd.read_csv(file)
+            df = pd.read_parquet(file)
             if "iter" not in df.columns:
                 print(f"Invalid file {file.name!r}, skipping.")
                 continue
@@ -53,3 +64,26 @@ def choose_file(results: List[Path]) -> Path:
             print(f"Invalid choice {choice!r}, exiting.")
             exit()
         return results[int(choice) - 1]
+
+
+def get_classification_datasets() -> list[tuple[DatasetAll, str]]:
+    return [
+        (fetch_open_payments(), "open_payments"),
+        (fetch_drug_directory(), "drug_directory"),
+        (fetch_road_safety(), "road_safety"),
+        (fetch_midwest_survey(), "midwest_survey"),
+        (fetch_traffic_violations(), "traffic_violations"),
+    ]
+
+
+def get_regression_datasets() -> list[tuple[DatasetAll, str]]:
+    return [
+        (fetch_medical_charge(), "medical_charge"),
+        (fetch_employee_salaries(), "employee_salaries"),
+    ]
+
+
+def get_dataset(info: tuple[DatasetAll, str]) -> tuple[pd.DataFrame, pd.Series]:
+    y = info.y
+    X = info.X
+    return X, y
