@@ -1,19 +1,17 @@
-from typing import Dict, List, Literal, Optional
-from warnings import warn
+from typing import Literal
 
 import numpy as np
 import pandas as pd
-from sklearn import __version__ as sklearn_version
 from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.utils.validation import check_is_fitted
 
-from skrub._utils import check_input, parse_version
+from skrub._utils import check_input
 
 # Required for ignoring lines too long in the docstrings
 # flake8: noqa: E501
 
 
-WORD_TO_ALIAS: Dict[str, str] = {
+WORD_TO_ALIAS: dict[str, str] = {
     "year": "Y",
     "month": "M",
     "day": "D",
@@ -24,7 +22,7 @@ WORD_TO_ALIAS: Dict[str, str] = {
     "microsecond": "us",
     "nanosecond": "N",
 }
-TIME_LEVELS: List[str] = list(WORD_TO_ALIAS.keys())
+TIME_LEVELS: list[str] = list(WORD_TO_ALIAS.keys())
 AcceptedTimeValues = Literal[
     "year",
     "month",
@@ -103,11 +101,12 @@ class DatetimeEncoder(BaseEstimator, TransformerMixin):
 
     n_features_in_: int
     n_features_out_: int
-    features_per_column_: Dict[int, List[str]]
-    col_names_: Optional[List[str]]
+    features_per_column_: dict[int, list[str]]
+    col_names_: list[str] | None
 
     def __init__(
         self,
+        *,
         extract_until: AcceptedTimeValues = "hour",
         add_day_of_the_week: bool = False,
     ):
@@ -229,7 +228,7 @@ class DatetimeEncoder(BaseEstimator, TransformerMixin):
 
         Returns
         -------
-        :obj:`~numpy.ndarray`, shape (n_samples, n_features_out_)
+        :obj:`~numpy.ndarray`, shape (n_samples, `n_features_out_`)
             Transformed input.
         """
         check_is_fitted(
@@ -253,7 +252,7 @@ class DatetimeEncoder(BaseEstimator, TransformerMixin):
             idx += len(self.features_per_column_[i])
         return X_
 
-    def get_feature_names_out(self, input_features=None) -> List[str]:
+    def get_feature_names_out(self, input_features=None) -> list[str]:
         """Return clean feature names.
 
         Feature names are formatted like: "<column_name>_<new_feature>"
@@ -278,28 +277,3 @@ class DatetimeEncoder(BaseEstimator, TransformerMixin):
             for feature in self.features_per_column_[i]:
                 feature_names.append(f"{prefix}_{feature}")
         return feature_names
-
-    def get_feature_names(self, input_features=None) -> List[str]:
-        """Return clean feature names. Compatibility method for sklearn < 1.0.
-
-        Use :func:`~DatetimeEncoder.get_feature_names_out` instead.
-
-        Parameters
-        ----------
-        input_features : None
-            Unused, only here for compatibility.
-
-        Returns
-        -------
-        list of str
-            List of feature names.
-        """
-        if parse_version(sklearn_version) >= parse_version("1.0"):
-            warn(
-                "Following the changes in scikit-learn 1.0, "
-                "get_feature_names is deprecated. "
-                "Use get_feature_names_out instead.",
-                DeprecationWarning,
-                stacklevel=2,
-            )
-        return self.get_feature_names_out()
