@@ -1,15 +1,10 @@
 import collections
-from typing import Any, Hashable
+from collections.abc import Hashable
+from typing import Any
 
 import numpy as np
+from sklearn.utils import parse_version  # noqa
 from sklearn.utils import check_array
-
-try:
-    # Works for sklearn >= 1.0
-    from sklearn.utils import parse_version  # noqa
-except ImportError:
-    # Works for sklearn < 1.0
-    from sklearn.utils.fixes import _parse_version as parse_version  # noqa
 
 
 class LRUDict:
