@@ -5,18 +5,17 @@ Functions that generate example data.
 from __future__ import annotations
 
 import string
-from typing import List, Optional, Union
 
 import numpy as np
 from sklearn.utils import check_random_state
 
 
 def make_deduplication_data(
-    examples: List[str],
-    entries_per_example: List[int],
+    examples: list[str],
+    entries_per_example: list[int],
     prob_mistake_per_letter: float,
-    random_state: Optional[Union[int, np.random.RandomState]] = None,
-) -> List[str]:
+    random_state: int | np.random.RandomState | None = None,
+) -> list[str]:
     """Duplicates examples with spelling mistakes.
 
     Characters are misspelled with probability `prob_mistake_per_letter`.
