@@ -1,7 +1,6 @@
 """
 Get the Wikipedia embeddings for feature augmentation.
 """
-from typing import Optional, Set
 
 import pandas as pd
 from sklearn.decomposition import PCA
@@ -18,7 +17,7 @@ _correspondence_table_url = (
 )
 
 
-def get_ken_table_aliases() -> Set[str]:
+def fetch_ken_table_aliases() -> set[str]:
     """Get the supported aliases of embedded KEN entities tables.
 
     These aliases can be using in subsequent functions (see section *See Also*).
@@ -30,28 +29,35 @@ def get_ken_table_aliases() -> Set[str]:
 
     See Also
     --------
-    :func:`get_ken_types`
+    :func:`fetch_ken_types`
         Helper function to search for entity types.
-    :func:`get_ken_embeddings`
+    :func:`fetch_ken_embeddings`
         Download Wikipedia embeddings by type.
 
     Notes
     -----
     Requires an Internet connection to work.
+
+    Examples
+    --------
+    Let's see what are the current KEN subtables available
+    for download:
+    >>> fetch_ken_table_aliases()
+    {'games', 'companies', 'schools', 'albums', 'all_entities', 'movies'}
     """
     correspondence = pd.read_csv(_correspondence_table_url)
     return set(["all_entities"] + list(correspondence["table"].values))
 
 
-def get_ken_types(
+def fetch_ken_types(
     search: str = None,
     *,
-    exclude: Optional[str] = None,
+    exclude: str | None = None,
     embedding_table_id: str = "all_entities",
 ) -> pd.DataFrame:
     """Helper function to search for KEN entity types.
 
-    The result can then be used with :func:`get_ken_embeddings`.
+    The result can then be used with :func:`fetch_ken_embeddings`.
 
     Parameters
     ----------
@@ -61,7 +67,7 @@ def get_ken_types(
         Substring pattern to exclude from the search.
     embedding_table_id : str, default='all_entities'
         Table of embedded entities from which to extract the embeddings.
-        Get the supported tables with :func:`get_ken_table_aliases`.
+        Get the supported tables with :func:`fetch_ken_table_aliases`.
         It is NOT possible to pass a custom figshare ID.
 
     Returns
@@ -71,7 +77,7 @@ def get_ken_types(
 
     See Also
     --------
-    :func:`get_ken_embeddings`
+    :func:`fetch_ken_embeddings`
         Download Wikipedia embeddings by type.
 
     References
@@ -82,10 +88,34 @@ def get_ken_types(
 
     Notes
     -----
-    Best used in conjunction with :func:`get_ken_embeddings`.
+    Best used in conjunction with :func:`fetch_ken_embeddings`.
+
+    Examples
+    --------
+    To get all the existing KEN types of entities:
+
+    >>> embedding_types = fetch_ken_types()
+    >>> embedding_types.head()
+                                Type
+    0                 wikicat_italian_male_screenwriters
+    1  wikicat_21st-century_roman_catholic_archbishop...
+    2                 wikicat_2000s_romantic_drama_films
+    3                  wikicat_music_festivals_in_france
+    4        wikicat_20th-century_american_women_artists
+
+    Let's search for all KEN types with the strings "dance" or "music":
+
+    >>> embedding_filtered_types = fetch_ken_types(search="dance|music")
+    >>> embedding_filtered_types.head()
+                                    Type
+    0                    wikicat_music_festivals_in_france
+    1  wikicat_films_scored_by_bharadwaj_(music_direc...
+    2                  wikicat_english_music_journalists
+    3       wikicat_20th-century_american_male_musicians
+    4  wikicat_alumni_of_the_london_academy_of_music_...
     """
     correspondence = pd.read_csv(_correspondence_table_url)
-    if embedding_table_id not in get_ken_table_aliases():
+    if embedding_table_id not in fetch_ken_table_aliases():
         raise ValueError(
             f"The embedding_table_id must be one of {correspondence['table'].unique()}."
         )
@@ -99,16 +129,17 @@ def get_ken_types(
         search_result = unique_types.X[unique_types.X["Type"].str.contains(search)]
     if exclude is not None:
         search_result = search_result[~search_result["Type"].str.contains(exclude)]
-    return search_result
+    search_result["Type"] = search_result["Type"].str[1:-1]
+    return search_result.reset_index(drop=True)
 
 
-def get_ken_embeddings(
-    types: Optional[str] = None,
+def fetch_ken_embeddings(
+    search_types: str | None = None,
     *,
-    exclude: Optional[str] = None,
+    exclude: str | None = None,
     embedding_table_id: str = "all_entities",
-    embedding_type_id: Optional[str] = None,
-    pca_components: Optional[int] = None,
+    embedding_type_id: str | None = None,
+    pca_components: int | None = None,
     suffix: str = "",
 ) -> pd.DataFrame:
     """Download Wikipedia embeddings by type.
@@ -118,7 +149,7 @@ def get_ken_embeddings(
 
     Parameters
     ----------
-    types : str, optional
+    search_types : str, optional
         Substring pattern that filters the types of entities.
         Will keep all entity types containing the substring.
         Write in lowercase. If `None`, all types will be passed.
@@ -126,11 +157,11 @@ def get_ken_embeddings(
         Type of embeddings to exclude from the types search.
     embedding_table_id : str, default='all_entities'
         Table of embedded entities from which to extract the embeddings.
-        Get the supported tables with :func:`get_ken_table_aliases`.
+        Get the supported tables with :func:`fetch_ken_table_aliases`.
         It is also possible to pass a custom figshare ID.
     embedding_type_id : str, optional
         Figshare ID of the file containing the type of embeddings.
-        Get the supported tables with :func:`get_ken_types`.
+        Get the supported tables with :func:`fetch_ken_types`.
         Ignored unless a custom `embedding_table_id` is provided.
     pca_components : int, optional
         Size of the dimensional space on which the embeddings will be projected
@@ -146,9 +177,9 @@ def get_ken_embeddings(
 
     See Also
     --------
-    :func:`get_ken_table_aliases`
+    :func:`fetch_ken_table_aliases`
         Get the supported aliases of embedded entities tables.
-    :func:`get_ken_types`
+    :func:`fetch_ken_types`
         Helper function to search for entity types.
     :func:`skrub.fuzzy_join` :
         Join two tables (dataframes) based on approximate column matching.
@@ -167,18 +198,51 @@ def get_ken_embeddings(
     The files are read and returned in parquet format, this function needs
     pyarrow installed to run correctly.
 
-    The `types` parameter is there to filter the types by the input string
+    The `search_types` parameter is there to filter the types by the input string
     pattern.
     In case the input is "music", all types with this string will be included
     (e.g. "wikicat_musician_from_france", "wikicat_music_label" etc.).
     Going directly for the exact type name (e.g. "wikicat_rock_music_bands")
     is possible but may not be complete (as some relevant bands may be
     in other similar types).
-    For searching the types, the :func:`~skrub.datasets.get_ken_types`
+    For exploring available types, the :func:`~skrub.datasets.fetch_ken_types`
     function can be used.
 
+    Examples
+    --------
+    :func:`fetch_ken_embeddings` allows you to extract embeddings
+    you are interested in. For instance, if we are interested in
+    video games:
+
+    >>> games_embedding = fetch_ken_embeddings(search_types="video_games")
+    >>> games_embedding.head()
+                    Entity                       Type                       ...      X198      X199
+    0       The_Mysterious_Island  wikicat_novels_adapted_into_video_games  ... -0.072814 -0.156973
+    1    Storyteller_(video_game)             wikicat_upcoming_video_games  ...  0.059816  0.021077
+    2  Skull_&_Bones_(video_game)        wikicat_video_games_about_pirates  ... -0.141682  0.024204
+    3               Ethan_Winters   wikicat_male_characters_in_video_games  ... -0.107913 -0.089531
+    4                     Cruis'n               wikicat_racing_video_games  ... -0.260757  0.060700
+
+    Extracts all embeddings with the "games" type.
+    For the list of existing types see :func:`fetch_ken_types`.
+
+    Some tables are available pre-filtered for us using the
+    `embedding_table_id` parameter:
+
+    >>> games_embedding_fast = fetch_ken_embeddings(embedding_table_id="games")
+    >>> games_embedding_fast.head()
+                     Entity                               Type                      ...      X198      X199
+    0              R-Type_Delta                                 wikicat_irem_games  ... -0.125806  0.040006
+    1  Just_Add_Water_(company)  wikicat_video_game_companies_of_the_united_kin...  ...  0.067210 -0.025676
+    2                 Li_Xiayan          wikicat_asian_games_medalists_in_swimming  ... -0.104818  0.003485
+    3             Vampire_Night                        wikicat_vampire_video_games  ... -0.118209 -0.145383
+    4               Shatterhand                             wikicat_platform_games  ... -0.138462  0.197820
+
+    It takes less time to load the wanted output, and is more precise as the
+    types have been carefully filtered out.
+    For a list of pre-filtered tables, see func:`fetch_ken_table_aliases`.
     """
-    if embedding_table_id in get_ken_table_aliases():
+    if embedding_table_id in fetch_ken_table_aliases():
         correspondence = pd.read_csv(_correspondence_table_url)
         embeddings_id = correspondence[correspondence["table"] == embedding_table_id][
             "entities_figshare_id"
@@ -189,8 +253,8 @@ def get_ken_embeddings(
     else:
         embeddings_id = embedding_table_id
     emb_type = fetch_figshare(embedding_type_id).X
-    if types is not None:
-        emb_type = emb_type[emb_type["Type"].str.contains(types)]
+    if search_types is not None:
+        emb_type = emb_type[emb_type["Type"].str.contains(search_types)]
     if exclude is not None:
         emb_type = emb_type[~emb_type["Type"].str.contains(exclude)]
     emb_type.drop_duplicates(subset=["Entity"], inplace=True)
@@ -217,4 +281,6 @@ def get_ken_embeddings(
         else:
             emb_final.append(emb_extracts)
     emb_df = pd.concat(emb_final)
-    return emb_df
+    emb_df["Entity"] = emb_df["Entity"].str[1:-1]
+    emb_df["Type"] = emb_df["Type"].str[1:-1]
+    return emb_df.reset_index(drop=True)
