@@ -1,9 +1,8 @@
 import os
 from pathlib import Path
-from typing import Optional
 
 
-def get_data_dir(name: Optional[str] = None) -> Path:
+def get_data_dir(name: str | None = None) -> Path:
     """
     Returns the directory in which skrub looks for data.
 
