@@ -1,3 +1,3 @@
 from ._argparser import default_parser
-from ._various import choose_file, find_result, find_results
+from ._various import choose_file, find_result, find_results, get_classification_datasets, get_regression_datasets, get_dataset
 from .monitor import monitor, repr_func
