@@ -122,11 +122,7 @@ encoder = make_column_transformer(
 #
 # We will use a |HGBR|,
 # which is a good predictor for data with heterogeneous columns
-# (we need to require the experimental feature for scikit-learn versions
-# earlier than 1.0):
-from sklearn.experimental import enable_hist_gradient_boosting
 
-# We can now import the |HGBR| from ensemble
 from sklearn.ensemble import HistGradientBoostingRegressor
 
 # We then create a pipeline chaining our encoders to a learner
@@ -217,7 +213,7 @@ plt.tight_layout()
 # data than the |MinHash| and |Gap|. The most scalable encoder is
 # the |MinHash|. On the other hand, the |Gap| has the benefit of
 # providing interpretable features
-# (see :ref:`sphx_glr_auto_examples_02_investigating_skrubegories.py`)
+# (see :ref:`sphx_glr_auto_examples_02_investigating_dirty_categories.py`)
 #
 # |
 #
