@@ -89,36 +89,40 @@ y = y[~mask]
 # We will use KEN embeddings to enrich our data.
 #
 # We will start by checking out the available tables with
-# :class:`~skrub.datasets.get_ken_table_aliases`:
-from skrub.datasets import get_ken_table_aliases
+# :class:`~skrub.datasets.fetch_ken_table_aliases`:
+from skrub.datasets import fetch_ken_table_aliases
 
-get_ken_table_aliases()
+fetch_ken_table_aliases()
 
 ###############################################################################
 # The *games* table is the most relevant to our case.
 # Let's see what kind of types we can find in it with the function
-# :class:`~skrub.datasets.get_ken_types`:
-from skrub.datasets import get_ken_types
+# :class:`~skrub.datasets.fetch_ken_types`:
+from skrub.datasets import fetch_ken_types
 
-get_ken_types(embedding_table_id="games")
+fetch_ken_types(embedding_table_id="games")
 
 ###############################################################################
 # Interesting, we have a broad range of topics!
 #
-# Next, we'll use :class:`~skrub.datasets.get_ken_embeddings`
+# Next, we'll use :class:`~skrub.datasets.fetch_ken_embeddings`
 # to extract the embeddings of entities we need:
-from skrub.datasets import get_ken_embeddings
+from skrub.datasets import fetch_ken_embeddings
 
 ###############################################################################
 # KEN Embeddings are classified by types.
-# The :class:`~skrub.datasets.get_ken_embeddings` function
+# See the example on :class:`~skrub.datasets.fetch_ken_embeddings`
+# to understand how you can filter types you are interested in.
+#
+# The :class:`~skrub.datasets.fetch_ken_embeddings` function
 # allows us to specify the types to be included and/or excluded
 # so as not to load all Wikipedia entity embeddings in a table.
 #
+#
 # In a first table, we include all embeddings with the type name "game"
 # and exclude those with type name "companies" or "developer".
-embedding_games = get_ken_embeddings(
-    types="game",
+embedding_games = fetch_ken_embeddings(
+    search_types="game",
     exclude="companies|developer",
     embedding_table_id="games",
 )
@@ -126,8 +130,8 @@ embedding_games = get_ken_embeddings(
 ###############################################################################
 # In a second table, we include all embeddings containing the type name
 # "game_development_companies", "game_companies" or "game_publish":
-embedding_publisher = get_ken_embeddings(
-    types="game_development_companies|game_companies|game_publish",
+embedding_publisher = fetch_ken_embeddings(
+    search_types="game_development_companies|game_companies|game_publish",
     embedding_table_id="games",
     suffix="_aux",
 )
