@@ -13,7 +13,7 @@ from ._fetching import (
     get_data_dir,
 )
 from ._generating import make_deduplication_data
-from ._ken_embeddings import get_ken_embeddings, get_ken_table_aliases, get_ken_types
+from ._ken_embeddings import fetch_ken_embeddings, fetch_ken_table_aliases, fetch_ken_types
 
 __all__ = [
     "DatasetAll",
@@ -29,7 +29,7 @@ __all__ = [
     "fetch_figshare",
     "get_data_dir",
     "make_deduplication_data",
-    "get_ken_embeddings",
-    "get_ken_table_aliases",
-    "get_ken_types",
+    "fetch_ken_embeddings",
+    "fetch_ken_table_aliases",
+    "fetch_ken_types",
 ]
