@@ -9,9 +9,9 @@ def test_get_data_dir(mock_os_path_dirname):
     """
     from skrub.datasets._utils import get_data_dir
 
-    expected_return_value_default = Path("/user/directory/data")
+    expected_return_value_default = Path("/user/directory/data").absolute()
 
-    mock_os_path_dirname.return_value = "/user/directory/"
+    mock_os_path_dirname.return_value = expected_return_value_default.parent
     assert get_data_dir() == expected_return_value_default
 
     expected_return_value_custom = expected_return_value_default / "tests"
