@@ -1,12 +1,10 @@
 import numpy as np
 import pandas as pd
 import pytest
-from sklearn import __version__ as sklearn_version
 from sklearn.exceptions import NotFittedError
 from sklearn.model_selection import train_test_split
 
 from skrub import GapEncoder
-from skrub._utils import parse_version
 from skrub.datasets import fetch_midwest_survey
 from skrub.tests.utils import generate_data
 
@@ -168,12 +166,7 @@ def test_get_feature_names_out(n_samples=70) -> None:
     X = generate_data(n_samples, random_state=0)
     enc = GapEncoder(random_state=42)
     enc.fit(X)
-    # Expect a warning if sklearn >= 1.0
-    if parse_version(sklearn_version) < parse_version("1.0"):
-        feature_names_1 = enc.get_feature_names()
-    else:
-        with pytest.warns(DeprecationWarning):
-            feature_names_1 = enc.get_feature_names()
+    feature_names_1 = enc.get_feature_names_out()
     feature_names_2 = enc.get_feature_names_out()
     for topic_labels in [feature_names_1, feature_names_2]:
         # Check number of labels
